feat: implement actions component with copy and feedback buttons
